modified:   client/index.html 	new file:   client/public/vlms_logo.png 	modified:   client/src/components/ui/user/Header.jsx 	deleted:    client/src/pages/user/About.jsx 	deleted:    client/src/pages/user/Browse.jsx 	renamed:    client/src/pages/user/MyList.jsx -> client/src/pages/user/Watchlist.jsx 	modified:   client/src/routes/App.jsx 	modified:   server/server.js
