+++ client/src/pages/user/Watchlist.jsx
